requirements.txt
gevent
requests
validators
python_telegram_bot>=5.3
flask
hjson
SQLAlchemy
